+++ src/utils/features.ts
@@ -9,6 +9,7 @@ const supportedFeatures = {
   shortUrlVisitsDeletion: { minVersion: '3.6.0' },
   orphanVisitsDeletion: { minVersion: '3.7.0' },
   shortUrlRedirectRules: { minVersion: '4.0.0' },
+  qrCodeColors: { minVersion: '4.0.0' },
   urlValidation: { maxVersion: '3.*.*' },
   ipRedirectCondition: { minVersion: '4.2.*' },
 } as const satisfies Record<string, Versions>;
@@ -18,7 +19,9 @@ Object.freeze(supportedFeatures);
 export type Feature = keyof typeof supportedFeatures;
 
 const isFeatureEnabledForVersion = (feature: Feature, serverVersion: SemVerOrLatest): boolean =>
-  serverVersion === 'latest' || versionMatch(serverVersion, supportedFeatures[feature]);
+  // When serverVersion is `latest`, fall back to a very big version number.
+  // That will disable features with a maxVersion, and keep enabled those with only a minVersion
+  versionMatch(serverVersion === 'latest' ? '999.99.99' : serverVersion, supportedFeatures[feature]);
 
 const getFeaturesForVersion = (serverVersion: SemVerOrLatest): Record<Feature, boolean> => ({
   excludeBotsOnShortUrls: isFeatureEnabledForVersion('excludeBotsOnShortUrls', serverVersion),
@@ -27,6 +30,7 @@ const getFeaturesForVersion = (serverVersion: SemVerOrLatest): Record<Feature, b
   shortUrlVisitsDeletion: isFeatureEnabledForVersion('shortUrlVisitsDeletion', serverVersion),
   orphanVisitsDeletion: isFeatureEnabledForVersion('orphanVisitsDeletion', serverVersion),
   shortUrlRedirectRules: isFeatureEnabledForVersion('shortUrlRedirectRules', serverVersion),
+  qrCodeColors: isFeatureEnabledForVersion('qrCodeColors', serverVersion),
   urlValidation: isFeatureEnabledForVersion('urlValidation', serverVersion),
   ipRedirectCondition: isFeatureEnabledForVersion('ipRedirectCondition', serverVersion),
 });