+++ test/short-urls/helpers/QrCodeModal.test.tsx
@@ -1,6 +1,7 @@
-import { fireEvent, screen, waitFor } from '@testing-library/react';
+import { fireEvent, screen } from '@testing-library/react';
 import { fromPartial } from '@total-typescript/shoehorn';
 import { QrCodeModalFactory } from '../../../src/short-urls/helpers/QrCodeModal';
+import { FeaturesProvider } from '../../../src/utils/features';
 import type { ImageDownloader } from '../../../src/utils/services/ImageDownloader';
 import { checkAccessibility } from '../../__helpers__/accessibility';
 import { renderWithEvents } from '../../__helpers__/setUpTest';
@@ -11,15 +12,20 @@ describe('<QrCodeModal />', () => {
     ImageDownloader: fromPartial<ImageDownloader>({ saveImage }),
   }));
   const shortUrl = 'https://s.test/abc123';
-  const setUp = () => renderWithEvents(
-    <QrCodeModal
-      isOpen
-      shortUrl={fromPartial({ shortUrl })}
-      toggle={() => {}}
-    />,
+  const setUp = ({ qrCodeColors = false }: { qrCodeColors?: boolean } = {}) => renderWithEvents(
+    <FeaturesProvider value={fromPartial({ qrCodeColors })}>
+      <QrCodeModal
+        isOpen
+        shortUrl={fromPartial({ shortUrl })}
+        toggle={() => {}}
+      />
+    </FeaturesProvider>,
   );
 
-  it('passes a11y checks', () => checkAccessibility(setUp()));
+  it.each([{ qrCodeColors: false }, { qrCodeColors: true }])(
+    'passes a11y checks',
+    ({ qrCodeColors }) => checkAccessibility(setUp({ qrCodeColors })),
+  );
 
   it('shows an external link to the URL in the header', () => {
     setUp();
@@ -31,23 +37,14 @@ describe('<QrCodeModal />', () => {
   });
 
   it('displays an image with the QR code of the URL', async () => {
-    setUp();
+    const { user } = setUp({ qrCodeColors: true });
+    const assertUrl = (url: string) => {
+      expect(screen.getByRole('img')).toHaveAttribute('src', url);
+      expect(screen.getByText(url)).toHaveAttribute('href', url);
+    };
 
-    const expectedUrl = `${shortUrl}/qr-code`;
-
-    expect(screen.getByRole('img')).toHaveAttribute('src', expectedUrl);
-    expect(screen.getByText(expectedUrl)).toHaveAttribute('href', expectedUrl);
-  });
-
-  it.each([
-    [530, 0, 'lg'],
-    [200, 0, undefined],
-    [830, 0, 'xl'],
-    [430, 80, 'lg'],
-    [200, 50, undefined],
-    [720, 100, 'xl'],
-  ])('renders expected size', async (size, margin, modalSize) => {
-    const { user } = setUp();
+    // Initially, no arguments are appended
+    assertUrl(`${shortUrl}/qr-code`);
 
     // Switch to sliders
     await user.click(screen.getByRole('button', { name: 'Customize size' }));
@@ -58,29 +55,30 @@ describe('<QrCodeModal />', () => {
       throw new Error('Sliders not found');
     }
 
-    fireEvent.change(sizeInput, { target: { value: `${size}` } });
-    fireEvent.change(marginInput, { target: { value: `${margin}` } });
+    // Setting size and margin should reflect in the QR code URL
+    fireEvent.change(sizeInput, { target: { value: '560' } });
+    fireEvent.change(marginInput, { target: { value: '20' } });
+    assertUrl(`${shortUrl}/qr-code?size=560&margin=20`);
 
-    expect(screen.getByText(`size: ${size}px`)).toBeInTheDocument();
-    expect(screen.getByText(`margin: ${margin}px`)).toBeInTheDocument();
+    // Unset margin and select error correction
+    await user.click(screen.getByRole('button', { name: 'Default margin' }));
+    await user.click(screen.getByRole('button', { name: 'Default error correction' }));
+    await user.click(screen.getByRole('menuitem', { name: /uartile/ }));
+    assertUrl(`${shortUrl}/qr-code?size=560&errorCorrection=Q`);
 
-    // Fake the images load event with a width that matches the size+margin
-    const image = screen.getByAltText('QR code');
-    Object.defineProperty(image, 'naturalWidth', {
-      get: () => size + margin,
-    });
-    image.dispatchEvent(new Event('load'));
-
-    if (modalSize) {
-      await waitFor(() => expect(screen.getByRole('document')).toHaveClass(`modal-${modalSize}`));
-    }
+    // Set custom color
+    await user.click(screen.getByRole('button', { name: 'Customize color' }));
+    assertUrl(`${shortUrl}/qr-code?size=560&errorCorrection=Q&color=000000`);
   });
 
-  it('shows expected buttons', () => {
-    setUp();
+  it.each([
+    { qrCodeColors: false, expectedButtons: 4 },
+    { qrCodeColors: true, expectedButtons: 6 },
+  ])('shows expected buttons', ({ expectedButtons, qrCodeColors }) => {
+    setUp({ qrCodeColors });
 
     // Add three because of the close, download and copy-to-clipboard buttons
-    expect(screen.getAllByRole('button')).toHaveLength(4 + 3);
+    expect(screen.getAllByRole('button')).toHaveLength(expectedButtons + 3);
   });
 
   it('saves the QR code image when clicking the Download button', async () => {